+++ Web/Program.cs
@@ -44,14 +44,12 @@ builder.Services
         s.Title = "FastEndpoints Sandbox";
         s.Version = "v2.0";
     })
-    .AddSwaggerDoc(maxEndpointVersion: 3, minEndpointVersion: 3, settings: s =>
+    .AddSwaggerDoc(minEndpointVersion: 3, maxEndpointVersion: 3, settings: s =>
     {
         s.DocumentName = "Release 3.0";
         s.Title = "FastEndpoints Sandbox ver3 only";
         s.Version = "v3.0";
     });
-
-
 
 var app = builder.Build();